requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
psycopg2==2.8.5
pycodestyle==2.6.0
pytz==2020.1
shortuuid==1.0.1
sqlparse==0.3.1
toml==0.10.1
